@@ -2,7 +2,7 @@ name: flutter_base_project
 description: A new Flutter project.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 1.0.2+3
+version: 1.0.3+4
 
 environment:
   flutter: 3.7.10
@@ -14,27 +14,29 @@ dependencies:
   flutter_localizations:
     sdk: flutter
 
-  shared_preferences: ^2.3.2
-  url_launcher: ^6.3.0
-
+  url_launcher: any # PLATFORM ANDROID IOS LINUX MACOS WEB WINDOWS
+  flutter_inappwebview: ^6.1.5
 
   # Network
-  http: ^1.2.2
-  connectivity: ^3.0.6
+  http: ^1.3.0 # PLATFORM ANDROID IOS LINUX MACOS WEB WINDOWS
+  connectivity_plus: ^6.1.3
   android_id: ^0.4.0
-  webview_flutter: ^4.8.0
+  webview_flutter: ^4.10.0 # PLATFORM ANDROID IOS
   webview_flutter_wkwebview: 3.17.0 # webview_flutter tarafından kullanılmakta ama son versiyon 3.18.0 olanda ios ta uygulamayı patlattığı için bu 3.17.0 versiyon sabitlendi daha sonra daha yeni versiyonlara geçilebilir
+  permission_handler: ^12.0.0+1
 
   # UI
   cupertino_icons: ^1.0.8
-  flutter_svg: ^2.0.10+1
-  flutter_local_notifications: ^17.2.2
-  banner_image: ^1.0.7
-  overlay_kit: ^1.0.8
+  flutter_svg: ^2.0.17 # PLATFORM ANDROID IOS LINUX MACOS WEB WINDOWS
+  flutter_local_notifications: ^19.1.0 #PLATFORM ANDROID IOS LINUX MACOS
+  banner_image: ^1.0.9 # PLATFORM ANDROID IOS MACOS WEB
+  overlay_kit: ^1.1.2 # PLATFORM ANDROID IOS LINUX MACOS WEB WINDOWS
+  path_provider: ^2.0.15
+  open_file: ^3.5.10
 
 
   # State manager
-  get: ^4.6.6
+  get: ^4.7.2 # PLATFORM ANDROID IOS LINUX MACOS WEB WINDOWS
 
   # Error handler
   dop_logger:
@@ -61,7 +63,6 @@ flutter_gen:
 flutter:
   uses-material-design: true
   assets:
-    #- assets/other/
 
 
   fonts:
