@@ -1,13 +1,17 @@
-import 'package:connectivity/connectivity.dart';
+import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter/material.dart' show BuildContext, FocusScope;
 
 extension CheckInternetExtension on BuildContext {
   ///it written for check the internet
   Future<bool> checkInternet() async {
-    var connectivityResult = await Connectivity().checkConnectivity();
-    if (connectivityResult == ConnectivityResult.none) {
-      return false;
-    } else {
+    try {
+      final connectivityResult = await Connectivity().checkConnectivity();
+      if (connectivityResult.first == ConnectivityResult.none) {
+        return false;
+      } else {
+        return true;
+      }
+    } catch (_) {
       return true;
     }
   }
