@@ -1,5 +1,2 @@
-import 'package:flutter_test/flutter_test.dart';
 
-void main() {
-  testWidgets('Counter increments smoke test', (WidgetTester tester) async {});
-}
+void main() {}
